+++ lib/galaxy/tools/parser/__init__.py
@@ -3,5 +3,8 @@
 from .interface import ToolSource
 from .factory import get_tool_source
 from .factory import get_input_source
+from .output_objects import (
+    ToolOutputCollectionPart,
+)
 
-__all__ = ["ToolSource", "get_tool_source", "get_input_source"]
+__all__ = ["ToolSource", "get_tool_source", "get_input_source", "ToolOutputCollectionPart"]
